// src/api/models/schedule.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Eternal Anvil (Backend)
 * The backend server for the Eternal Anvil.
 * OpenAPI spec version: v0.0.1
 */

export interface Schedule {
  friday?: number[];
  monday?: number[];
  saturday?: number[];
  sunday?: number[];
  thursday?: number[];
  tuesday?: number[];
  wednesday?: number[];
}